--- js/utils/choose.js
@@ -1,4 +1,0 @@
-export function choose(choices) {
-  const index = Math.floor(Math.random() * choices.length);
-  return choices[index];
-}
