--- js/utils/getCommentNumberMargin.js
@@ -1,1 +1,0 @@
-export const getCommentNumberMargin = (i) => (i - 1) * 5;
