--- js/utils/getRandomDiceNumber.js
@@ -1,11 +1,0 @@
-//Функция, возвращающая случайное целое число из переданного диапазона
-
-export const getRandomDiceNumber = (min, max) => {
-  if (min < 0 || max < 0) {
-    return -1;
-  }
-  if (min > max) {
-    [min, max] = [max, min];
-  }
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-};
